# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Online Library API with CRUD operations.",
        "title": "Online Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/paidbook": {
            "get": {
                "description": "Get list of all paidbooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paidbook"
                ],
                "summary": "Get all paidbooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/paidbook.PaidBook"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new paidbook with file upload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paidbook"
                ],
                "summary": "Create new paidbook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Book Summary",
                        "name": "summary",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Author",
                        "name": "author",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Book Pages",
                        "name": "pages",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Book Price",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Cover Image",
                        "name": "cover_image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PDF File",
                        "name": "pdf_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/paidbook/{id}": {
            "get": {
                "description": "Get a paidbook by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paidbook"
                ],
                "summary": "Get paidbook by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/paidbook.PaidBook"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update paidbook details by ID (with optional new files)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paidbook"
                ],
                "summary": "Update paidbook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Book Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Summary",
                        "name": "summary",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Book Author",
                        "name": "author",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Book Pages",
                        "name": "pages",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Book Price",
                        "name": "price",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Cover Image",
                        "name": "cover_image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "PDF File",
                        "name": "pdf_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a paidbook by ID",
                "tags": [
                    "paidbook"
                ],
                "summary": "Delete paidbook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "paidbook.PaidBook": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover_image": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "pdf_file": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}